requirements.txt
beautifulsoup4==4.14.2
langchain==0.3.27
langchain_community==0.3.31
langchain_core==0.3.79
langchain_huggingface==0.3.1
langchain_openai==0.3.35
langchain_text_splitters==0.3.11
langsmith==0.4.37
python-dotenv==1.1.1
Requests==2.32.5
rich==14.2.0
